+++ drivers/iommu/io-pgtable.h
@@ -157,7 +157,7 @@ struct io_pgtable_ops {
 	int (*map_sg)(struct io_pgtable_ops *ops, unsigned long iova,
 		      struct scatterlist *sg, unsigned int nents,
 		      int prot, size_t *size);
-	size_t (*unmap)(struct io_pgtable_ops *ops, unsigned long iova,
+	int (*unmap)(struct io_pgtable_ops *ops, unsigned long iova,
 		     size_t size);
 	phys_addr_t (*iova_to_phys)(struct io_pgtable_ops *ops,
 				    unsigned long iova);
